Make parameter separator comparison portable
Compare C characters after lossless promotion to i32 so the code compiles when c_char is unsigned on aarch64 and signed on x86_64.

+++ rclrs/src/parameter/service.rs
@@ -133,7 +133,14 @@ fn list_parameters(req: ListParameters_Request, map: &ParameterMap) -> ListParam
         if req.depth == ListParameters_Request::DEPTH_RECURSIVE {
             return true;
         }
-        u64::try_from(substring.iter().filter(|c| **c == b'.').count()).unwrap() < req.depth
+        u64::try_from(
+            substring
+                .iter()
+                .filter(|c| i32::from(**c) == i32::from(b'.'))
+                .count(),
+        )
+        .unwrap()
+            < req.depth
     };
     let names: Sequence<_> = map
         .storage
